+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
-    tools:context="network.o3.o3wallet.LoginActivity">
+    tools:context="network.o3.o3wallet.Onboarding.LoginActivity">
 
     <TextView
         android:id="@+id/textView3"
