+++ app/src/main/res/layout/fragment_landing_feature_scroll.xml
@@ -4,7 +4,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="250dp"
-    tools:context="network.o3.o3wallet.LandingFeatureScroll">
+    tools:context="network.o3.o3wallet.Onboarding.LandingFeatureScroll">
 
     <!-- TODO: Update blank fragment layout -->
